+++ src/base/b-dynamic-page/interface.ts
@@ -72,8 +72,3 @@ export interface OnBeforeSwitchPage {
 	 */
 	saveScroll(el: Element): void;
 }
-
-type PageComponentName = string;
-type PageKey = string;
-
-export type PageInfo = [PageComponentName, PageKey?];