+++ src/base/b-dynamic-page/CHANGELOG.md
@@ -9,11 +9,11 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
-## v3.75.0 (2024-07-28)
+## v3.74.4 (2024-08-02)
 
-#### :boom: Breaking Change
+#### :rocket: New Feature
 
-* The `page` and `eventConverter` props now accept arrays consisting of the page component name and page component key, instead of just a string representing the page component name
+* The `eventConverter` function can now return a tuple consisting of the page component name and page component key, instead of just a string representing the page component name
 
 #### :bug: Bug Fix